+++ ui/litellm-dashboard/eslint-suppressions.json
@@ -66,17 +66,17 @@
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/add_margin_form.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/add_margin_form.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/add_provider_form.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/add_provider_form.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/cost_tracking_settings.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/cost_tracking_settings.tsx": {
     "no-nested-ternary": {
       "count": 2
     },
@@ -84,62 +84,62 @@
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/how_it_works.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/how_it_works.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/pricing_calculator/multi_cost_results.test.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/pricing_calculator/multi_cost_results.test.tsx": {
     "unused-imports/no-unused-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/pricing_calculator/multi_cost_results.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/pricing_calculator/multi_cost_results.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/pricing_calculator/multi_export_dropdown.test.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/pricing_calculator/multi_export_dropdown.test.tsx": {
     "unused-imports/no-unused-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/pricing_calculator/multi_export_dropdown.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/pricing_calculator/multi_export_dropdown.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/pricing_calculator/use_multi_cost_estimate.ts": {
+  "src/app/(dashboard)/cost-tracking/_components/pricing_calculator/use_multi_cost_estimate.ts": {
     "no-restricted-syntax": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/provider_discount_table.test.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/provider_discount_table.test.tsx": {
     "unused-imports/no-unused-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/provider_discount_table.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/provider_discount_table.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/provider_display_helpers.test.ts": {
+  "src/app/(dashboard)/cost-tracking/_components/provider_display_helpers.test.ts": {
     "unused-imports/no-unused-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/provider_margin_table.tsx": {
+  "src/app/(dashboard)/cost-tracking/_components/provider_margin_table.tsx": {
     "no-restricted-imports": {
       "count": 1
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/use_discount_config.ts": {
+  "src/app/(dashboard)/cost-tracking/_components/use_discount_config.ts": {
     "no-restricted-syntax": {
       "count": 2
     }
   },
-  "src/app/(dashboard)/cost-tracking/components/use_margin_config.ts": {
+  "src/app/(dashboard)/cost-tracking/_components/use_margin_config.ts": {
     "no-restricted-syntax": {
       "count": 2
     }
